requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.18.0
assemblyai>=0.17.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
isort>=5.12.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.0.0
